requirements.txt
Django==2.0.7
django-debug-toolbar==1.9.1
django-rest-framework==0.1.0
djangorestframework==3.8.2
pytz==2018.5
sqlparse==0.2.4
